feat(onboarding)!: the conversation IS the onboarding — wizard deleted
Phase 6: the conversational shell renders unconditionally and the flag,
the stepper coordinator, the Footer, the form-era VoiceStep and the
wizard wrappers are gone (onboarding.tsx: 3,210 to 354 lines, with the
surviving shared pieces split into named modules: company form, manual
interview, results card, connect panels). Every invariant the 18
pinned wizard tests protected either ports to the conversational
surface or already had its equivalent there - the mapping is in the PR.
75 dead i18n keys removed per locale. OAuth-return deep links ride the
conversational restore unchanged. STATUS.md records the milestone.
