+++ frontend/src/i18n/en.ts
@@ -823,24 +823,13 @@ export const en = {
   "stage.semWon": "Won",
   "stage.semLost": "Lost",
 
-  "ob.title": "Set up Margince",
   "ob.read": "Read",
   "ob.confirm": "Confirm",
-  "ob.company": "Company",
-  "ob.voice": "Voice",
-  "ob.results": "Results",
-  "ob.connect": "Connect",
   "ob.url": "Website",
   "ob.urlScheme": "https://",
-  "ob.next": "Continue",
   "ob.back": "Back",
   "ob.finish": "Into the workspace",
-  "ob.skipStep": "Skip this step",
   "ob.restoring": "Restoring your setup…",
-  "ob.stateConflict":
-    "This setup changed in another tab. We reloaded the saved version so nothing is silently overwritten.",
-  "ob.stateSaveFailed":
-    "We couldn't save your setup progress. Please try again.",
   "ob.readKick": "Step 1 of 4 · company context",
   "ob.readTitle": "Your company",
   "ob.readSub": "Read it from your website, or type it in.",
@@ -979,21 +968,10 @@ export const en = {
   "ob.s1.omitBody":
     "I filled only what I could quote from your site. You can add anything I couldn't verify.",
   "ob.readGo": "Read my website",
-  "ob.reading": "Reading…",
-  "ob.readAgain": "Read again",
   "ob.trustPublic": "I only read your public website. No login required.",
   "ob.urlWillRead": "I'll read {host}",
-  "ob.urlInvalid":
-    "That doesn't look like a website address — try e.g. {example}",
   "ob.readFromSite": "read from site",
   "ob.failTitle": "I couldn't read enough from this website",
-  "ob.failBody":
-    "I reached the site but couldn't find enough public text to ground a profile, so I won't guess. I saved nothing.",
-  "ob.failWhy1":
-    "The page may render its content with JavaScript we can't read from the public HTML",
-  "ob.failWhy2":
-    "It may be a login wall, a holding page, or block automated reads in robots.txt",
-  "ob.failWhy3": "There may simply be too little text to work from",
   "ob.tryAnother": "Try another URL",
 
   "ob.manualChapterLegal": "Your legal organization",
@@ -1071,76 +1049,6 @@ export const en = {
   "ob.field.industry": "Industry",
   "ob.field.history": "Company history",
 
-  "ob.s2.kick": "Step 2 of 4 · optional",
-  "ob.s2.title": "Now let's learn",
-  "ob.s2.titleEm": "how you really write.",
-  "ob.s2.sub":
-    "We don't guess your tone from two emails. Feed us a real corpus — posts, emails, and especially call transcripts — and we model your actual voice: cadence, signature moves, the words you use and the ones you never do. Aim for ~30,000 words. Everything stays in your workspace.",
-  "ob.s2.optinTitle": "Why now — and why it's optional.",
-  "ob.s2.optinBody":
-    "You've seen your business read back (Step 1) and you'll see a working draft next (Step 3). This step makes those drafts sound unmistakably like you. It's the one step that asks for your own content, so it's entirely your call.",
-  "ob.s2.optinYes": "Yes, build my voice now",
-  "ob.s2.optinSkip": "Skip — I'll do this later in Settings",
-  "ob.s2.dropTitle": "Drag & drop files here, or browse",
-  "ob.s2.dropFmt":
-    ".txt .md .vtt .srt .json · text only · we keep only your turns",
-  "ob.s2.dropSkipped":
-    "Skipped {files} — not a text format. Convert to .txt/.md or paste the text instead.",
-  "ob.s2.speakerAsk": "Which speaker are you in {file}?",
-  "ob.s2.speakerOption": "{name} · words: {words} · turns: {turns}",
-  "ob.s2.keptOnly": "counted {kept} of {total} — only what {speaker} said",
-  "ob.s2.unattributed":
-    "I can't tell who says what in {file}, so none of it is counted. Add speaker labels, or paste only your own words.",
-  "ob.reg.spoken": "spoken",
-  "ob.reg.written": "written",
-  "ob.reg.casual": "casual",
-  "ob.src.emails": "Sent emails",
-  "ob.src.emailsHint":
-    "Your richest voice source for email drafting — unlocks when you connect your inbox",
-  "ob.src.transcripts": "Call / meeting transcripts",
-  "ob.src.transcriptsHint":
-    "Fathom, Otter, Fireflies — we keep only your turns",
-  "ob.src.posts": "LinkedIn / X posts",
-  "ob.src.postsHint": "Your post history — export or paste",
-  "ob.src.longform": "Blog posts / newsletters",
-  "ob.src.longformHint": "Substack, Medium, Notion exports",
-  "ob.src.chat": "Slack / WhatsApp",
-  "ob.src.chatHint": "Messages you sent",
-  "ob.src.memos": "Voice memos",
-  "ob.src.memosHint": "Record a 2-min monologue — transcribed",
-  "ob.s2.whenConnected": "learns from this once connected",
-  "ob.s2.words": "{count} / {target} words",
-  "ob.s2.qualStart": "add a few sources to start",
-  "ob.s2.qualThin": "thin — add a transcript or two",
-  "ob.s2.qualGood": "good — getting there",
-  "ob.s2.qualRich": "rich — a distinctive voice takes shape",
-  "ob.s2.qualSharp": "sharp — unmistakably you",
-  "ob.s2.mix": "spoken {spoken}% · written {written}% · {sources} sources",
-  "ob.s2.spokenHint":
-    "A handful of call transcripts beats 30 LinkedIn posts — spoken sources carry your real cadence. For transcripts we keep only your turns.",
-  "ob.s2.emailCallout":
-    "Your sent emails are your richest voice source. This first Voice DNA gets much sharper the moment you connect your inbox (the last step), then keeps learning from every email you send.",
-  "ob.s2.pasteLabel": "Or paste writing samples",
-  "ob.s2.pastePlaceholder":
-    "Paste a few emails, posts, or anything you've written in your own words…",
-  "ob.s2.pasteSource": "Pasted writing",
-  "ob.s2.build": "Build my Voice DNA",
-  "ob.s2.building": "Building your Voice DNA from {count} words…",
-  "ob.s2.minWords":
-    "Add at least {min} words in your own voice to build a first Voice DNA.",
-  "ob.s2.vpMeta": "{count} words · {sources} sources",
-  "ob.s2.builtTitle": "Your Voice DNA",
-  "ob.s2.builtEmpty":
-    "Your Voice DNA is building — it'll appear here and in Settings shortly.",
-  "ob.s2.deferredTitle": "Building your Voice DNA",
-  "ob.s2.deferredBody":
-    "It's queued and will be ready shortly — you can move on. It sharpens automatically once you connect your inbox in the last step.",
-  "ob.s2.failedTitle": "We couldn't build it just now",
-  "ob.s2.failedBody":
-    "Something went wrong building your Voice DNA. Try again, or skip and build it later in Settings.",
-  "ob.s2.vpFootnote":
-    "Built from {count} words. Connect your inbox next and your sent email pushes this sharper — then it keeps learning from every email you send.",
-
   "ob.s3.kick": "Step 3 of 4",
   "ob.s3.title": "Look what you've built —",
   "ob.s3.titleEm": "with nothing connected.",
@@ -1173,11 +1081,7 @@ export const en = {
     "No magic — it's the standard B2B stage template, tuned to your industry from the Step-1 read. It's empty right now. When you connect your inbox, capture reads your sent mail and meetings, then proposes deals into these stages — each one evidenced and reversible. You approve what becomes a deal.",
   "ob.s3.stillNothing":
     "Still nothing connected. You're in control of when that changes.",
-  "ob.s3.cta": "Connect my inbox",
 
-  "ob.s4.kick": "Step 4 of 4 · the last step",
-  "ob.s4.title": "Ready to",
-  "ob.s4.titleEm": "bring it to life?",
   "ob.s4.sub":
     "You've built the brain. Connect your inbox and your CRM fills itself — people, companies, and deals captured automatically, so you never type them.",
   "ob.s4.provGoogle": "Google",
@@ -1786,8 +1690,6 @@ export const en = {
   "voice.outcome.rollback": "restored",
   "voice.history.versionRow": "v{n} \u00b7",
   "voice.history.loadMore": "Show older entries",
-  "ob.s2.candidateNote":
-    "This version is waiting for your review \u2014 apply or reject it under Settings \u2192 Voice.",
   "voice.insights.provenance": "Built from your corpus \u00b7 v{n}",
   "voice.insights.statWords": "Words: {count}",
   "voice.insights.statSources": "Sources: {count}",
